pyproject.toml
[tool.poetry]
name = "elect"
version = "0.1.0"
description = ""
authors = ["Lotos <shelkovnikova.da@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
load-dotenv = "^0.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
